// eur-example/switchboard-function/sbv32/devops/alerting-service/src/networks/types.ts
export type ChainType = (typeof SWITCHBOARD_CHAINS)[number];
// ...
type IBaseQueueConfig = {
  name: string;
  address: string;
  settings?: Record<string, JsonType>;
};

type ISwitchboardAttestationQueueConfig = IBaseQueueConfig;

type ISwitchboardQueueConfig = IBaseQueueConfig & {
  crankAddress?: string;
  permissioned: boolean;
};
// ...
type INetworkMetadata = {
  [key: string]: JsonType | undefined;
  defaultRpcUrl: string;
};

type IBaseNetworkConfig = {
  chain: ChainType;
  networkName: string;
  address: string;
  metadata: INetworkMetadata;
};

export type IEvmNetworkConfig = IBaseNetworkConfig & {
  sbPushOracle: string;
  chainId: number;
  queues: Array<ISwitchboardQueueConfig>;
  attestationQueues: Array<ISwitchboardAttestationQueueConfig>;
};

export type IStarknetNetworkConfig = IBaseNetworkConfig & {
  sbPushOracle: string;
  chainId: string;
  queues: Array<ISwitchboardQueueConfig>;
  attestationQueues: Array<ISwitchboardAttestationQueueConfig>;
};

export type IChainNetworkConfig = IBaseNetworkConfig & {
  attestationService?: string;
  queues: Array<ISwitchboardQueueConfig>;
  attestationQueues: Array<ISwitchboardAttestationQueueConfig>;
};
// ...
export type IChainConfig = {
  mainnet: IChainNetworkConfig;
  testnet: IChainNetworkConfig;
};